Symlink node and npm before attempting to install npm packages
Previously, yt.formula.node would install node and then attempt to
install npm packages before symlinking the node and npm binaries to the
bin directory (which is in PATH). This causes npm package installation
to fail as npm can't be invoked before the node symlink is present.

This change resolves this issue by updating the installation order to:
- Install node and npm
- Symlink the node and npm executables to the bin directory (in PATH)
- Install npm packages